fix headless game mode
