+++ tuxemon/main.py
@@ -32,10 +32,8 @@ of the game.
 """
 
 import sys
-import pygame
 from core.main import main
 
 if __name__ == '__main__':
     main()
-    pygame.quit()
     sys.exit()
